server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
livekit-api==1.0.7
python-dotenv==1.2.1

